Backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
google-generativeai==0.8.4
langchain==0.3.21
langchain-community==0.3.20
pyjwt==2.10.1
werkzeug==3.1.3
boto3==1.35.0
botocore==1.35.0
